+++ s2core/src/main/scala/com/kakao/s2graph/core/Util.scala
@@ -4,7 +4,7 @@ import java.util.concurrent.atomic.AtomicBoolean
 
 import com.google.common.cache.CacheBuilder
 import play.api.Logger
-import play.api.libs.json.JsValue
+import play.api.libs.json.{JsString, JsNumber, JsValue}
 
 import scala.concurrent.{Future, ExecutionContext}
 import scala.util.{Success, Failure}
@@ -95,5 +95,28 @@ class SafeUpdateCache[T, M[_]](prefix: String, maxSize: Int, ttl: Int)(implicit 
   }
 }
 
+class MultiOrdering(ascendingLs: Seq[Boolean], defaultAscending: Boolean = true) extends Ordering[Seq[JsValue]] {
+  override def compare(x: Seq[JsValue], y: Seq[JsValue]): Int = {
+    val xe = x.iterator
+    val ye = y.iterator
+    val oe = ascendingLs.iterator
 
+    while (xe.hasNext && ye.hasNext) {
+      val ascending = if (oe.hasNext) oe.next() else defaultAscending
+      val (xev, yev) = ascending match {
+        case true => xe.next() -> ye.next()
+        case false => ye.next() -> xe.next()
+      }
+      val res = (xev, yev) match {
+        case (JsNumber(xv), JsNumber(yv)) =>
+          Ordering.BigDecimal.compare(xv, yv)
+        case (JsString(xv), JsString(yv)) =>
+          Ordering.String.compare(xv, yv)
+        case _ => throw new Exception(s"unsupported type")
+      }
+      if (res != 0) return res
+    }
 
+    Ordering.Boolean.compare(xe.hasNext, ye.hasNext)
+  }
+}
